Move `MobilizonEvent.compute_status` to `storage.query`
